1. Add English Support; 2. Separate choice into a table;
Why separat choices into a single table?
The description will be used in questions and reports table. Only store
choices in questions table is not enough. But for this, since for each
questionnaire query, there will be a join among three tables. This may be
a performance issue. Will try and modify if needed.